Updated functions/s2_scl_weights.js (added 1 line, deleted 1 line).

+++ functions/s2_scl_weights.js
@@ -10,7 +10,7 @@ exports.s2_scl_weights = function(img_coll, AOI, scale_to_use){
         reducer: ee.Reducer.mode(),
         geometry: region.geometry(),
         scale: scale_to_use
-      }).getNumber("SCL").toByte();
+      }).getNumber("SCL");
       
       return ee.Image(scl_value).clipToBoundsAndScale({
         geometry: region.geometry(),
